+++ CMS/ASN1/PKCS8/PKCS8_Input.h
@@ -21,9 +21,9 @@ public:
 	void	Get_PrivateKeyInfo(void);
 	void	Get_PrivateKey(
 				unsigned	char*		_key,
-				unsigned	int			_szKey);
+							size_t		_szKey);
 	void	Get_PrivateKey_with_check(
 				AlgorithmIdentifier*	_algorithm,
 				unsigned	char*		_key,
-				unsigned	int			_szKey);
+							size_t		_szKey);
 };
